+++ __init__.py
@@ -158,6 +158,7 @@ class EclipseExtension(ComfyExtension):
         from .py.RvImage_TextImageWithFX import RvImage_TextImageWithFX
         from .py.RvImage_ImageWithFX import RvImage_ImageWithFX
         from .py.RvImage_AlignSize import RvImage_AlignSize
+        from .py.RvImage_GetLastImage import RvImage_GetLastImage
         # Loader nodes
         from .py.legacy.legacy_Checkpoint_Loader_Small import RvLoader_Checkpoint_Loader_Small
         from .py.legacy.legacy_Checkpoint_Loader_Small_Pipe import RvLoader_Checkpoint_Loader_Small_Pipe
@@ -362,6 +363,7 @@ class EclipseExtension(ComfyExtension):
             RvImage_TextImageWithFX,
             RvImage_ImageWithFX,
             RvImage_AlignSize,
+            RvImage_GetLastImage,
             # Loader
             RvLoader_Checkpoint_Loader_Small,
             RvLoader_Checkpoint_Loader_Small_Pipe,